Select dialog seems to work
Not much testing yet.

refers to #55
